Import of NATG/Class-AutoDB-1.29_03 from CPAN.
gitpan-cpan-distribution: Class-AutoDB
gitpan-cpan-version:      1.29_03
gitpan-cpan-path:         NATG/Class-AutoDB-1.29_03.tar.gz
gitpan-cpan-author:       NATG
gitpan-cpan-maturity:     developer

+++ t/autodb.999.cleanup.t
@@ -12,10 +12,13 @@ my $ok;
 my $file=File::Spec->catfile(qw(t testdb));
 if (-e $file) {
   my $testdb=testdb;
-  my $dbh=DBI->connect("dbi:mysql:",$ENV{USER},undef,
-		       {AutoCommit=>1, ChopBlanks=>1, PrintError=>0, PrintWarn=>0, Warn=>0,});
-  $ok=$dbh->do(qq(DROP DATABASE IF EXISTS $testdb));
-  ok($ok,"drop database $testdb");
+  if ($testdb ne 'test') {
+    # 'test' database is pre-existing. don't drop
+    my $dbh=DBI->connect("dbi:mysql:",$ENV{USER},undef,
+			 {AutoCommit=>1, ChopBlanks=>1, PrintError=>0, PrintWarn=>0, Warn=>0,});
+    $ok=$dbh->do(qq(DROP DATABASE IF EXISTS $testdb));
+    ok($ok,"drop database $testdb");
+  }
   $ok=unlink($file);
   ok($ok,"delete testdb file");
 }
